Initial commit.
`Dockerfile`s for Sagittariidae base and application containers.

These containers have been created to expedite an imminent deployment.
They're not strictly correct as they build in-place and include non-prod
resources.  Ideally, I believe that built artefacts should be pulled
into the containers that are going to be used for deployment.

For now, however, these are sufficient to get an alpha service deployed.