Package version updates.
Updating package versions to work around `apk` update failures, because,
horifically, it is not possible to pin package versions in Alpine.

There is a [conversation][forum] in the Alpine Linux forum that has more
detail.  Unfortunately the suggested workaround of pinning the release
version doesn't work in this case, so all we can do is update the
package versions.

This makes repeatable builds impossible and is a compelling reason to
move (back) to Ubuntu, despite it's larger image size.

[forum]: http://forum.alpinelinux.org/forum/general-discussion/it-not-possible-install-old-version-packages